[experimental] mk_event_libevent: re-attach pipe fds from provided ev_map
This seems to resolve the issue with constant handle and socket leak,
however, on the other hand - it uncovers other issue, where one-shot
event has not been properly destroyed, as with this change, we can see
that its timeout callback is fired, without any harm.

Leaking resource is a Linux-style pipe emulation provided by libevent,
used by Fluent-Bit asynchronous internals for scenarios like async
DNS query, etc.

NOTE: This is an experimental change, not a fix.